# docs/frontend-backend/openapi-snapshot.yaml
openapi: 3.1.0
info:
  title: BigHead API
  version: 0.3.0
  description: 'Contrato canônico Sprint 3: matriz T01-T56 e implementação FastAPI.'
servers:
- url: http://localhost:8000
  description: Local
paths:
...
  /v1/crm/connections:
    get:
      tags:
      - crm-integrations
      summary: List Connections
      operationId: crmConnectionsGet
      security:
      - HTTPBearer: []
      parameters:
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CrmConnection'
                title: Response Crmconnectionsget
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - crm-integrations
      summary: Create Connection
      operationId: crmConnectionsPost
      security:
      - HTTPBearer: []
      parameters:
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrmConnectionCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrmConnection'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/crm/connections/{connection_id}/sync:
    post:
      tags:
      - crm-integrations
      summary: Request Sync
      operationId: crmConnectionSyncPost
      security:
      - HTTPBearer: []
      parameters:
      - name: connection_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Connection Id
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Crmconnectionsyncpost
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
...
  schemas:
...
    CrmConnection:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        providerKey:
          type: string
          title: Providerkey
        displayName:
          type: string
          title: Displayname
        status:
          type: string
          title: Status
        configuration:
          additionalProperties: true
          type: object
          title: Configuration
        lastSyncedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Lastsyncedat
      type: object
      required:
      - id
      - providerKey
      - displayName
      - status
      - configuration
      title: CrmConnection
    CrmConnectionCreate:
      properties:
        providerKey:
          type: string
          pattern: ^[a-z0-9][a-z0-9_-]{1,62}$
          title: Providerkey
        displayName:
          type: string
          maxLength: 200
          minLength: 1
          title: Displayname
        configuration:
          additionalProperties: true
          type: object
          title: Configuration
      type: object
      required:
      - providerKey
      - displayName
      title: CrmConnectionCreate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...